# backend/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Swagger API Docs for Loan Management System",
      "description": "This is a sample Loan Management aAPI documentation based on the OpenAPI 3.0 specification.",
      "version": "1.0.11"
    },
    "servers": [
      {
        "url": "/"
      }
    ],
    "schemes": ["http", "https"],
...
    "paths": {
...
      "/api/users/login": {
        "post": {
          "tags": ["User"],
          "summary": "Login user",
          "description": "Authenticate an existing user.",
          "operationId": "loginUser",
          "requestBody": {
            "description": "Created user object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userLogin"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/userLogin"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/userLogin"
                }
              }
            }
          },
          "responses": {
            "default": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "application/xml": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "User": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "example": "theuser@gmail.com"
            },
            "username": {
              "type": "string"
            }
          },
          "json": {
            "name": "User"
          }
        },
...
        "userLogin": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "json": {
            "name": "userLogin"
          }
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        },
        "api_key": {
          "type": "apiKey",
          "name": "api_key",
          "in": "header"
        }
      }
    }
  }